requirements.txt
Django==3.1.7
django-admin==2.0.1
djangorestframework==3.12.4